core/requirements.txt
Django==5.2.5
djangorestframework==3.14.0
psycopg2-binary==2.9.7
fastapi==0.95.2
uvicorn==0.22.0
